requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile requirements.in
#
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
google-api-core==2.3.2
    # via google-api-python-client
google-api-python-client==2.33.0
    # via -r requirements.in
google-auth==2.3.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==0.4.6
    # via -r requirements.in
googleapis-common-protos==1.54.0
    # via google-api-core
httplib2==0.20.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
oauthlib==3.1.1
    # via requests-oauthlib
protobuf==3.19.1
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.6
    # via httplib2
requests==2.26.0
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.7
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
